+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.dev
 
 * FIX: `pytoil init` now gracefully exits on ctrl+c preventing a bug where if a user aborted halfway through, the config file would appear to have been written but would not be valid and raise an ugly error on next use.
+* NEW: All commands that were in `pytoil project` are now available from the root command. E.g. instead of typing `pytoil project checkout` you may now simply use `pytoil checkout`.
 
 ## 0.1.0
